Adding main menu music
By using a 'play sound 2D' node I was able to implement the game music without having to create an unnecessary 'play sound' variable. this would be able to bypass the need for that and immediately take the sound file from the content drawer. In order for the main menu music to loop again if the players stays on the main menu i added the retriggerable delay which links to the 'play sound 2D' node